requirements.txt
asgiref==3.8.1
boto3==1.35.2
botocore==1.35.2
Django==4.2.15
image==1.5.33
jmespath==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.19
